+++ Plugin.Jobs/CrossJobs.cs
@@ -5,8 +5,24 @@ namespace Plugin.Jobs
 {
     public static partial class CrossJobs
     {
-        public static bool IsLoggingEnabled { get; set; } = true;
+        static CrossJobs()
+        { 
+            ResolveJob = (jobInfo) => 
+            {
+                if (jobInfo.Type == null)
+                    throw new ArgumentException($"Job '{jobInfo.Name}' class type not found");
 
+                var job = Activator.CreateInstance(jobInfo.Type) as IJob;
+                if (job == null)
+                    throw new ArgumentException("Type is not IJob or was not found");
+
+                return job;
+            };
+        }
+
+
+        public static bool IsLoggingEnabled { get; set; } = true;
+        public static Func<JobInfo, IJob> ResolveJob { get; set; }
 
         static IJobManager current;
         public static IJobManager Current