+++ Sample/Sample.Android/MainActivity.cs
@@ -24,7 +24,6 @@ namespace Sample.Droid
             TabLayoutResource = Resource.Layout.Tabbar;
             ToolbarResource = Resource.Layout.Toolbar;
             base.OnCreate(savedInstanceState);
-            CrossJobs.Init(this, savedInstanceState);
             UserDialogs.Init(() => (Activity)Forms.Context);
             Forms.Init(this, savedInstanceState);
             this.LoadApplication(new App());