+++ Sample/Sample/Jobs/SampleJob.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using System.Threading;
 using System.Threading.Tasks;
-using Acr.UserDialogs;
 using Plugin.Jobs;
 
 
@@ -9,16 +8,18 @@ namespace Sample.Jobs
 {
     public class SampleJob : IJob
     {
-        //readonly IUserDialogs dialogs;
-        //public SampleJob(IUserDialogs dialogs) => this.dialogs = dialogs;
-
+        readonly ISampleDependency dependency;
+        public SampleJob(ISampleDependency dependency)
+        {
+            this.dependency = dependency;
+        }
 
         public async Task Run(JobInfo jobInfo, CancellationToken cancelToken)
         {
             // you really shouldn't interact with the UI from a job
             //this.dialogs.Alert("HELLO FROM JOB");
             var loops = jobInfo.GetValue("LoopCount", 25);
-
+            
             for (var i = 0; i < loops; i++)
             {
                 if (cancelToken.IsCancellationRequested)